Make datadir-initializing functions more NFS-friendly.
Unless you mount with no_root_squash, creating dirs and
setting permissions inside an NFS share as root means
you're going to have a bad time.

Refactored to use a couple of convenience functions that
check for environment-based flags to see if they should
run these operations as the GITLAB_USER instead of as
root.

It's possible we might just want to run these as
GITLAB_USER all the time by default, since running things
in containers as root is bad form anyway.